+++ CHANGELOG.md
@@ -5,3 +5,4 @@ Flutter Project Setup CLI Changelog
 - Implemented flutter_project_setup_cli create command template with specified args
 - Implemented dynamic links, and JWT authentication setups
 - Implemented State Management setup with Riverpod
+- Implemented Notification setup with Firebase Cloud Messaging